+++ exp/static/exp/css/app.css
@@ -102,6 +102,15 @@ a.state-link {
 .pl-xl {
     padding-left: 20px;
 }
+.mb-xs {
+    margin-bottom:10px;
+}
+.mb-sm {
+    margin-bottom: 15px;
+}
+.mb-md {
+    margin-bottom: 20px;
+}
 .mb-lg {
     margin-bottom: 25px;
 }
